requirements.txt
omegaconf
argparse
easydict
matplotlib
numpy
opencv-python
scipy
torchvision
tensorboardX
